+++ src/Logger.php
@@ -8,14 +8,41 @@ use Psr\Log\InvalidArgumentException;
 
 class Logger extends AbstractLogger
 {
+    const DEFAULT_VERBOSITY = 6;
+
+    protected static $logLevels = array(
+        LogLevel::EMERGENCY,
+        LogLevel::ALERT,
+        LogLevel::CRITICAL,
+        LogLevel::ERROR,
+        LogLevel::WARNING,
+        LogLevel::NOTICE,
+        LogLevel::INFO,
+        LogLevel::DEBUG,
+    );
+
+    protected $verbosity;
+
+    public function __construct()
+    {
+        $this->verbosity = self::DEFAULT_VERBOSITY;
+    }
+
     public function log($level, $message, array $context = array())
     {
         if (!$this->isLogLevelKnown($level)) {
             throw new InvalidArgumentException("level $level is invalid");
         }
 
-        $message = $this->getMessage($level, $message, $context);
-        error_log($message);
+        if ($this->shouldLog($level)) {
+            $message = $this->getMessage($level, $message, $context);
+            fwrite(STDERR, $message . PHP_EOL);
+        }
+    }
+
+    public function setVerbosity($verbosity)
+    {
+        $this->verbosity = $verbosity;
     }
 
     protected function getMessage($level, $message, $context)
@@ -59,5 +86,10 @@ class Logger extends AbstractLogger
         }
 
         return strtr($message, $replace);
+    }
+
+    protected function shouldLog($level)
+    {
+        return array_search($level, self::$logLevels) < $this->verbosity;
     }
 }